requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
PyPDF2>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
chromadb>=0.4.0

# Phase 1: Report Output Enhancement
weasyprint>=60.0
matplotlib>=3.8.0

# Phase 2: Interactive Experience Improvements
tqdm>=4.66.0
rich>=13.7.0
click>=8.1.0

# Phase 3: Performance and Cache Optimization
tiktoken>=0.5.0
diskcache>=5.6.0